refactor: Regroup nextcloud queries inside same file as others
