+++ src/modules/nextcloud/helpers.js
@@ -1,0 +1,3 @@
+export const computeNextcloudFolderQueryId = ({ sourceAccount, path }) => {
+  return `io.cozy.remote.nextcloud.files/sourceAccount/${sourceAccount}/path${path}`
+}
